INT-7053: Organize library into ApolloV1/ApolloV2 folders; one type per v2 file
- Move the v1 sources under Sources/DataDomeApollo/ApolloV1/ and the v2 sources
  under ApolloV2/, with one type per file on the v2 side (DataDomeError,
  DataDomeInterceptor, DataDomeInterceptorProvider).
- Rename the v1 interceptor type DataDomeResponseInterceptor -> DataDomeInterceptor
  so both Apollo majors expose the same DataDomeInterceptor / DataDomeInterceptorProvider
  public API.
- SwiftPM requires unique source filenames within a target, so the two types present
  in both majors use a symmetric +ApolloVx filename qualifier
  (DataDomeInterceptor+ApolloV1.swift / +ApolloV2.swift, and the provider likewise);
  version-unique types (DataDomeRequestContext, DataDomeError) keep clean names.

Verified: both trait paths build (Apollo 1.25.6 and 2.1.2).
