+++ Sources/DataDomeApollo/ApolloV1/DataDomeInterceptor+ApolloV1.swift
@@ -1,8 +1,12 @@
 //
-//  DataDomeResponseInterceptor.swift
+//  DataDomeInterceptor+ApolloV1.swift
 //  DataDomeApollo
 //
 //  Created by Mohamed Hajlaoui on 31/03/2021.
+//
+//  Defines `DataDomeInterceptor` for Apollo iOS v1. The filename is qualified `+ApolloV1` because
+//  SwiftPM requires unique source file names within a target — the v2 definition of the same type
+//  lives in `ApolloV2/DataDomeInterceptor+ApolloV2.swift`.
 //
 
 #if ApolloV1
@@ -12,7 +16,7 @@ import ApolloAPI
 import CoreDataDome
 
 /// The DataDome interceptor. Use this to get your networking pipeline protected.
-public final class DataDomeResponseInterceptor: ApolloInterceptor {
+public final class DataDomeInterceptor: ApolloInterceptor {
     public let id: String = UUID().uuidString
 
     /// The CoreDataDome SDK instance used to validate responses.
