+++ Sources/DataDomeApollo/ApolloV2/DataDomeInterceptor+ApolloV2.swift
@@ -1,5 +1,5 @@
 //
-//  DataDomeApollo+ApolloV2.swift
+//  DataDomeInterceptor+ApolloV2.swift
 //  DataDomeApollo
 //
 //  Apollo iOS v2 support, built on the async `GraphQLInterceptor` API.
@@ -11,12 +11,6 @@ import Foundation
 import Apollo
 import ApolloAPI
 import CoreDataDome
-
-/// Errors surfaced by the DataDome Apollo v2 interceptor.
-public enum DataDomeError: Error {
-    /// DataDome presented a block page; the request cannot proceed.
-    case blocked
-}
 
 /// A CoreDataDome `GraphQLInterceptor` for Apollo iOS v2.
 ///
@@ -81,32 +75,6 @@ public struct DataDomeInterceptor: GraphQLInterceptor {
     private static func isDataDomeChallenge(_ response: HTTPURLResponse) -> Bool {
         response.value(forHTTPHeaderField: "x-dd-b") != nil
             || response.value(forHTTPHeaderField: "x-sf-cc-x-dd-b") != nil
-    }
-}
-
-/// An Apollo iOS v2 `InterceptorProvider` that appends ``DataDomeInterceptor`` to Apollo's default
-/// GraphQL interceptors.
-///
-/// DataDome is added last (innermost), so its `mapErrors` sees a `ResponseCodeError` before the other
-/// GraphQL interceptors, while `MaxRetryInterceptor` (first/outermost) still bounds the retry loop. All
-/// other interceptors — including the HTTP `ResponseCodeInterceptor` whose error carries the challenge
-/// response — use Apollo's defaults.
-public struct DataDomeInterceptorProvider: InterceptorProvider {
-
-    /// The CoreDataDome SDK instance used to validate responses.
-    private let dataDome: DataDome
-
-    /// Creates an interceptor provider backed by a CoreDataDome SDK instance.
-    /// - Parameter dataDome: The CoreDataDome SDK instance used to validate responses.
-    public init(dataDome: DataDome) {
-        self.dataDome = dataDome
-    }
-
-    public func graphQLInterceptors<Operation: GraphQLOperation>(
-        for operation: Operation
-    ) -> [any GraphQLInterceptor] {
-        DefaultInterceptorProvider.shared.graphQLInterceptors(for: operation)
-            + [DataDomeInterceptor(dataDome: dataDome)]
     }
 }
 #endif
